Hilariously, rotation causes the precise collision-detection to fall apart.
Did not previously take into account that serious trig. would be
required to account for the modified orientation of the edges in the
calculation.

This much-simpler and less-classy alternative approach “draws” a big
ol’ static hitbox for the spin attack, roughly tracing the
circumference of the attack’s tip. “Roughly,” because it’s a rectangle.

Oh, well.